--- backend/src/main/java/de/cae/XYFleet/Driver.java
@@ -1,7 +1,0 @@
-package de.cae.XYFleet;
-
-import org.restlet.resource.ServerResource;
-
-public class Driver extends XYServerResource {
-
-}
