--- backend/src/main/java/de/cae/XYFleet/User.java
@@ -1,6 +1,0 @@
-package de.cae.XYFleet;
-
-import org.restlet.resource.ServerResource;
-
-public class User extends XYServerResource {
-}
