+++ backend/src/main/java/de/cae/XYFleet/ressource/Driver.java
@@ -1,0 +1,5 @@
+package de.cae.XYFleet.ressource;
+
+public class Driver extends XYServerResource {
+
+}
