+++ backend/src/main/java/de/cae/XYFleet/ressource/User.java
@@ -1,0 +1,4 @@
+package de.cae.XYFleet.ressource;
+
+public class User extends XYServerResource {
+}
