+++ backend/src/main/java/de/cae/XYFleet/ressource/XYServerResource.java
@@ -1,5 +1,6 @@
-package de.cae.XYFleet;
+package de.cae.XYFleet.ressource;
 
+import de.cae.XYFleet.Database;
 import org.jooq.DSLContext;
 import org.jooq.JSONFormat;
 import org.jooq.Record;
@@ -8,14 +9,6 @@ import org.jooq.tools.json.JSONObject;
 import org.restlet.data.Status;
 import org.restlet.resource.ResourceException;
 import org.restlet.resource.ServerResource;
-import org.w3c.dom.Document;
-
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-
-import java.sql.ResultSet;
-
-import static de.cae.XYFleet.Database.initDatabase;
 
 public abstract class XYServerResource extends ServerResource {
